moved renderReplySuccess
